requirements.txt
pandas
numpy
pyyaml
scikit-learn
yfinance
pyarrow
matplotlib

torch
tqdm

mlflow
fastapi
uvicorn

feedparser
vaderSentiment

requests